pyproject.toml
[project]
name = "ms24-sum25-agentic-ai-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "langchain[google-genai]>=0.3.26",
    "langgraph>=0.5.1",
    "mammoth>=1.9.1",
    "pydantic>=2.11.7",
    "pytidylib>=0.3.2",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "ruff>=0.12.2",
]

[project.scripts]
main = "src.app:main"
